move params to compose, add them to docs, add keypoints utils to docs, fix all notebooks
