+++ docs/api/augmentations.rst
@@ -15,3 +15,8 @@ Helper functions for working with bounding boxes
 ------------------------------------------------
 .. automodule:: albumentations.augmentations.bbox_utils
     :members:
+
+Helper functions for working with keypoints
+------------------------------------------------
+.. automodule:: albumentations.augmentations.keypoints_utils
+    :members:
