+++ notebooks/example_bbox_keypoint_rotate.py
@@ -36,7 +36,8 @@ def main():
 
     aug = A.Compose([
         A.ShiftScaleRotate(scale_limit=0.1, shift_limit=0.2, rotate_limit=10, always_apply=True)
-    ], bbox_params={'format': 'pascal_voc', 'label_fields': ['bbox_labels']}, keypoint_params={'format': 'xy'})
+    ], bbox_params=A.BboxParams(format='pascal_voc', label_fields=['bbox_labels']),
+        keypoint_params=A.KeypointParams(format='xy'))
 
     for i in range(10):
         data = aug(image=image, keypoints=keypoints, bboxes=bboxes, bbox_labels=np.ones(len(bboxes)))
